git-svn-id: svn+ssh://svn.r-forge.r-project.org/svnroot/genabel/branches/ProbABEL-pacox/v.0.3.0/ProbABEL@1230 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ src/coxph_data.cpp
@@ -114,7 +114,7 @@ coxph_data::coxph_data(phedata &phed, gendata &gend, int snpnum)
     {
         for (int j = 0; j < ngpreds; j++)
         {
-            float snpdata[nids];
+            double snpdata[nids];
             gend.get_var(snpnum * ngpreds + j, snpdata);
             for (int i = 0; i < nids; i++)
                 X.put(snpdata[i], i, (ncov - ngpreds + j));
@@ -195,7 +195,7 @@ void coxph_data::update_snp(gendata &gend, int snpnum)
 
     for (int j = 0; j < ngpreds; j++)
     {
-        float snpdata[nids];
+        double snpdata[nids];
         for (int i = 0; i < nids; i++)
         {
             masked_data[i] = 0;
